pyproject.toml
[build-system]
requires = [
    "setuptools>=61.0",
    "wheel"
]
build-backend = "setuptools.build_meta"

[project]
name = "ai-data-synthesizer"
version = "0.1.0"
description = "Basic data science project template"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.10.0",
    "fastapi>=0.115.12",
    "google-cloud-logging>=3.11.4",
    # "fastapi[standard]>=0.115.11",
    # "httpx>=0.28.1",
    "langchain>=0.3.20",
    "langchain-experimental>=0.3.4",
    "langchain-google-firestore>=0.5.0",
    "langchain-google-vertexai>=2.0.14",
    "langchain-ollama>=0.2.3",
    "pandas>=2.2.3",
    "pypdf>=5.4.0",
    "streamlit>=1.43.2",
    "toml>=0.10.2",
    "tqdm>=4.67.1",
    #"unstructured[pdf]",
    "types-pyyaml==6.0.*",
    "types-toml==0.10.*",
    "uvicorn>=0.34.0",
    "watchdog>=6.0.0",
    "mypy>=1.15.0",
    "black>=25.1.0",
    "ipykernel>=6.29.5",
    "ragas>=0.2.14",
    "opik>=1.7.4",
    "langfuse==2.60.*",
    "deepeval==2.7.*",
    "chonkie[st]>=1.4.0",
    "openpyxl>=3.1.5",
]

authors = [
    { name = "Alexander Vosseler", email = "avosseler@nemetschek.com" }
]

keywords = ["ChatBot", "AI", "Assistant", "Nemy"]

[project.optional-dependencies]
dev = [
  "jupyter",
  "ipykernel",
#   "types-requests",
  "pytest",
#   "bandit",
#   "flake8",
#   "flake8-pyproject",
#   "isort",
#   "pytest-asyncio",
#   "pytest-mock",
  "pytest-timeout",
#   "pdoc",
#   "websocket-client",
#   "pre-commit"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["ai_eval*"]

[tool.mypy]
python_version = "3.12"
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.black]
line-length = 88
skip-string-normalization = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra"
testpaths = [
    "tests"
]
